requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.3.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
alembic==1.14.0
pytest==8.3.4
httpx==0.28.1
